@@ -2,13 +2,13 @@ buildscript {
 
     ext {
         kotlinVersion = '1.2.41'
-        androidGradlePluginVersion = '3.2.0-alpha15'
+        androidGradlePluginVersion = '3.2.0-alpha17'
         dokkaVersion = '0.9.17'
     }
 
     repositories {
-        jcenter()
         google()
+        jcenter()
     }
 
     dependencies {
@@ -17,23 +17,24 @@ buildscript {
         classpath 'com.github.dcendents:android-maven-gradle-plugin:2.1'
         classpath 'com.jfrog.bintray.gradle:gradle-bintray-plugin:1.8.0'
         classpath "org.jetbrains.dokka:dokka-android-gradle-plugin:$dokkaVersion"
-        classpath 'digital.wup:android-maven-publish:3.3.0'
     }
 }
 
 allprojects {
     repositories {
-        jcenter()
         google()
+        jcenter()
     }
 }
 
 ext {
-    compileSdkVersion = 27
-    buildToolsVersion = "27.0.3"
-    supportLibVersion = '27.1.1'
+    compileSdkVersion = 28
+    buildToolsVersion = "28.0.0"
     minSdkVersion = 14
-    targetSdkVersion = 27
+    targetSdkVersion = 28
+    appcompatVersion = '1.0.0-alpha3'
+    recyclerVersion = '1.0.0-alpha3'
+    lifecycleVersion = '2.0.0-alpha1'
 }
 
 task clean(type: Delete) {
