@@ -2,10 +2,9 @@ package com.otaliastudios.elements.sample
 
 import android.content.Context
 import android.os.Bundle
-import android.databinding.DataBindingUtil
-import android.support.v4.app.Fragment
-import android.support.v7.app.AppCompatActivity
-import android.support.v7.widget.RecyclerView
+import androidx.fragment.app.Fragment
+import androidx.appcompat.app.AppCompatActivity
+import androidx.recyclerview.widget.RecyclerView
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
@@ -20,7 +19,7 @@ class MainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.main_activity)
-        val recycler = findViewById<RecyclerView>(R.id.recycler)
+        val recycler = findViewById<androidx.recyclerview.widget.RecyclerView>(R.id.recycler)
         val list = listOf(R.string.menu_paged, R.string.menu_paged_on_click, R.string.menu_shuffle,
                 R.string.menu_headers, R.string.menu_empty, R.string.menu_complete)
         val presenter = BottomPresenter(this, ::onElementClick)
@@ -46,14 +45,14 @@ class MainActivity : AppCompatActivity() {
             else -> null
         }
         if (klass != null) {
-            val fragment = Fragment.instantiate(this, klass.java.name)
+            val fragment = androidx.fragment.app.Fragment.instantiate(this, klass.java.name)
             supportFragmentManager.beginTransaction()
                     .replace(R.id.fragmentContainer, fragment)
                     .commit()
         }
     }
 
-    abstract class BaseFragment : Fragment() {
+    abstract class BaseFragment : androidx.fragment.app.Fragment() {
 
         override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
             return inflater.inflate(R.layout.list, container, false)
@@ -61,15 +60,15 @@ class MainActivity : AppCompatActivity() {
 
         override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
             super.onViewCreated(view, savedInstanceState)
-            setUp(context!!, view as RecyclerView)
+            setUp(context!!, view as androidx.recyclerview.widget.RecyclerView)
         }
 
-        abstract fun setUp(context: Context, recyclerView: RecyclerView)
+        abstract fun setUp(context: Context, recyclerView: androidx.recyclerview.widget.RecyclerView)
     }
 
     class PagedFragment : BaseFragment() {
 
-        override fun setUp(context: Context, recyclerView: RecyclerView) {
+        override fun setUp(context: Context, recyclerView: androidx.recyclerview.widget.RecyclerView) {
             val message = "This list shows pages of 10 elements each. New pages are loaded when needed " +
                     "(that is, when you scroll down) depending on the pageSize value that is passed " +
                     "to the adapter constructor.\n\n" +
@@ -89,7 +88,7 @@ class MainActivity : AppCompatActivity() {
 
     class ClickPagedFragment : BaseFragment() {
 
-        override fun setUp(context: Context, recyclerView: RecyclerView) {
+        override fun setUp(context: Context, recyclerView: androidx.recyclerview.widget.RecyclerView) {
             val message = "The Adapter loads new pages based on the position of the requested item " +
                     "and the pageSize value that is passed to the constructor. By default it is Int.MAX_VALUE: this means " +
                     "that the page opening behavior is disabled, and it acts as a single page Adapter. " +
@@ -112,7 +111,7 @@ class MainActivity : AppCompatActivity() {
 
     class ShuffleFragment : BaseFragment() {
 
-        override fun setUp(context: Context, recyclerView: RecyclerView) {
+        override fun setUp(context: Context, recyclerView: androidx.recyclerview.widget.RecyclerView) {
             val message = "A single page list with some colors. Each 3 seconds, the list of colors is shuffled. " +
                     "It is extremely easy to send new results for the current page: they will replace the old results " +
                     "for this page coming from the same Source. The library will use DiffUtil " +
@@ -130,7 +129,7 @@ class MainActivity : AppCompatActivity() {
 
     class EmptyFragment : BaseFragment() {
 
-        override fun setUp(context: Context, recyclerView: RecyclerView) {
+        override fun setUp(context: Context, recyclerView: androidx.recyclerview.widget.RecyclerView) {
             val message = "This is what happens when a MainSource posts an empty list as a result. " +
                     "It will automatically pass an 'empty' element that you can present with a EmptyPresenter." +
                     "The same happens when such source posts an Exception. To handle this case, use an ErrorPresenter.\n\n" +
@@ -146,7 +145,7 @@ class MainActivity : AppCompatActivity() {
 
     class HeadersFragment : BaseFragment() {
 
-        override fun setUp(context: Context, recyclerView: RecyclerView) {
+        override fun setUp(context: Context, recyclerView: androidx.recyclerview.widget.RecyclerView) {
             val message = "An example of Source dependencies and ordering. " +
                     "We declare a CheeseSource for the main list and a separate CheeseHeaderSource for the alphabet headers. " +
                     "In the header class, we say that we depend on CheeseSource results, and compute the headers.\n\n" +
@@ -167,7 +166,7 @@ class MainActivity : AppCompatActivity() {
 
     class CompleteFragment : BaseFragment() {
 
-        override fun setUp(context: Context, recyclerView: RecyclerView) {
+        override fun setUp(context: Context, recyclerView: androidx.recyclerview.widget.RecyclerView) {
             val message = "All the previous features are expressed in Source and Presenter objects. This means " +
                     "that we can easily combine them by just adding them to the Adapter builder. " +
                     "Using addSource and addPresenter, we can compose all features in a single adapter, " +
